Implemented dynamic health bar

+++ Assets/Projectiles/Scripts/IDamageable.cs
@@ -5,12 +5,12 @@
         /// <summary>
         /// Objects hit points
         /// </summary>
-        public float HitPoint { get; set; }
+        public int HitPoint { get; set; }
         
         /// <summary>
         /// Retrieving the damage from a projectile  
         /// </summary>
         /// <param name="damage">Amount of damage taken</param>
-        public void ReceivedDamage(float damage);
+        public void ReceivedDamage(int damage);
     }
 }
